consolidate setup flows for both device types
