+++ packages/blockchain-wallet-v4-frontend/src/data/components/lockbox/actionTypes.js
@@ -53,6 +53,7 @@ export const INITIALIZE_DASHBOARD = '@LOCKBOX.INITIALIZE_DASHBOARD'
 export const UPDATE_TRANSACTION_LIST = '@LOCKBOX.UPDATE_TRANSACTION_LIST'
 
 // APPLICATIONS
+export const NEW_DEVICE_BTC_INSTALL = '@LOCKBOX.NEW_DEVICE_BTC_INSTALL'
 export const INITIALIZE_APP_MANAGER = '@LOCKBOX.INITIALIZE_APP_MANAGER'
 export const SET_LATEST_APP_INFOS_LOADING =
   '@LOCKBOX.SET_LATEST_APP_INFOS_LOADING'
@@ -66,8 +67,3 @@ export const APP_CHANGE_LOADING = '@LOCKBOX.APP_CHANGE_LOADING'
 export const APP_CHANGE_SUCCESS = '@LOCKBOX.APP_CHANGE_SUCCESS'
 export const APP_CHANGE_FAILURE = '@LOCKBOX.APP_CHANGE_FAILURE'
 export const RESET_APP_CHANGE_STATUS = '@LOCKBOX.RESET_APP_CHANGE_STATUS'
-
-// TODO: remove in LB2
-export const NEW_DEVICE_BTC_INSTALL = '@LOCKBOX.NEW_DEVICE_BTC_INSTALL'
-export const NEW_DEVICE_BTC_INSTALL_NEXT =
-  '@LOCKBOX.NEW_DEVICE_BTC_INSTALL_NEXT'
