+++ packages/blockchain-wallet-v4-frontend/src/modals/Lockbox/LockboxSetup/OpenBtcAppStep/template.js
@@ -2,13 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { FormattedMessage } from 'react-intl'
 import { RotateSync } from 'components/RotateSync'
-import {
-  Button,
-  Image,
-  Link,
-  Text,
-  TextGroup
-} from 'blockchain-info-components'
+import { Button, Image, Text, TextGroup } from 'blockchain-info-components'
 
 const Title = styled.div`
   text-align: center;
@@ -36,15 +30,9 @@ const ButtonContainer = styled.div`
 const RotateSyncContainer = styled(RotateSync)`
   margin-left: 15px;
 `
-const InstallTexts = styled(TextGroup)`
-  margin-top: 12px;
-  & > :last-child {
-    margin-left: -3px;
-  }
-`
 
 const OpenBtcAppStep = props => {
-  const { isReady, hideInstallLink, onInstallApps, onStepChange } = props
+  const { isReady, onStepChange } = props
 
   return (
     <React.Fragment>
@@ -69,28 +57,6 @@ const OpenBtcAppStep = props => {
             defaultMessage='Select &quot;Bitcoin&quot; from the picker.'
           />
         </Text>
-        {!hideInstallLink && (
-          <InstallTexts inline>
-            <Text size='14px' weight={400}>
-              <FormattedMessage
-                id='modals.lockboxsetup.openbtcappstep.note'
-                defaultMessage='Note:'
-              />
-            </Text>
-            <Text size='14px' weight={300}>
-              <FormattedMessage
-                id='modals.lockboxsetup.openbtcappstep.installapps'
-                defaultMessage='If you do not have the Bitcoin app on your device, you can install it'
-              />
-            </Text>
-            <Link size='14px' weight={400} onClick={onInstallApps}>
-              <FormattedMessage
-                id='modals.lockboxsetup.openbtcappstep.installappshere'
-                defaultMessage='here.'
-              />
-            </Link>
-          </InstallTexts>
-        )}
       </Content>
       <ImageContainer>
         <Image
